Fix dragging on unix
Previously, if the mouse went below 70px from the top of the window,
the dragging operation would stop. This was because of a check that was
put in place to prevent dragging by moving the mouse anywhere.

To fix this, a boolean was added which will be set to true on the first
click and drag on the 70px area. If the boolean is true, then it will
drag until the mouse is relased.